backend/requirements.txt
fastapi
uvicorn
web3
python-decouple
requests